core: significant rework to config subsystem
This change hammers out a bunch of details and changes the form of
the config from being a hashmap based on a primary name to just
being a list.  There is some gymnastics in order to properly parse
the toml to a form that has the deafults filled out and such.  This
should make using the resultant GpioConfig easier for other parts
of the system.

Signed-off-by: Paul Osborne <osbpau@gmail.com>