options: use &str over String where possible
I had been avoiding this as it means that we have to use
some explicit lifetimes, but all-in-all using explicit lifetimes
turned out not to be too bad and made the code overall a bit
cleaner and easier to understand.

Should see minor performance and memory usage gains from this
change.

Signed-off-by: Paul Osborne <osbpau@gmail.com>