Allow specifying two common options via env
Configuration files an symlink root are used in many subcommands and are
expected to be fixed for a certain environment. Allow to conveniently pass them
via environment variables.

The symlink root has a somewhat special role as it might as well be specified
via configuration files and we can't have a default value from the CLI in this
case.